Update main.html
adjust the content of create post pop up window because I combine the starting time and date, and ending time and date from 4 column to 2 column.
In addition, remove the hard code post dropdown collapse. try to create a render function to insert the post to the specific dvi container.